Many new features: ZML. Upload ability.
The community plugin is required, but its settings are simply ignored
by icy_picture_modify. If community is disabled, the upload function
will not work any more.

The code of the community 2.3.e is a bit buggy in Piwigo 2.4. Moreover,
its "moderate" feature is not pluggable. I have to copy another version
of the file "add_photos.php" and create some personal patch. Fortunately,
Piwigo plugin system is powerful enough to allow me to unload the file
from the community plugin =)) Cool !

In this feature, ACL is loaded from a ZML file in local_dir. With the new
feature `reference` it is easy to write various kind of permissions.
ZML is a name similar to YML ;) I do not want to have all features of YML.

I have spent a whole Saturday on the plugin. Today is a bad sad day.
No one prays for me.... and I am crying with coding.