Fix template issue: hide something as they should be hidden.
New method to check if any image is editable. In the page, we should hide
the button for sync,delete if the current user can not delete the image.

The sync_ method does not work any more in Piwigo 2.4. Not sure if they
are still working on Piwigo 2.3. Disable it completely on Piwigo 2.4