Fix a serious bug that breaks ACL. Thanks to Kalle
We use a wrong JOIN command (left join instead of a inner join),
that brings the highest permissions to a users. If there are some
groups, the user can always get the permissions from the highest
group. This is an effect of the use of LEFT-JOIN query.

If you are using version >= 2.1.0 of this plugin, it is highly
that you upgrade to the latest version. If you don"t want to upgrade,
you can

* Edit the file manually, by replacing the LEFT JOIN by JOIN.
  Please search and edit in the file
   plugins/icy_picture_modify/include/*.php
* Apply a patch found from the source tree
   https://github.com/icy/icy_picture_modify/tree/master/patches/
  (Find your version and patch file "IPM-SA-2013-08-14.patch"

I hate PHP :)