!bug: Windows CR/LF will break the parser
Due to strict regular expression the ZAML parser, and the $
does not match any linefeed, the author name is ignored,
and that is why any new setting will go to :default section.

We can fix this by change the regexp, or simplify :rtrim
the input $line. The 2nd way is much simpler, right?