!bug: show two Edit buttons for :admin (piwigo 2.5)
Piwigo < 2.5 does not have this problem: When an image is editable,
two Edit buttons will be displayed: One of IPM, the other is of
native Piwigo. To fix this issue (remove our button), we will
check if the current user is *admin. If *yes, just exit. Because
we can not detect user information at the time the hooks are loaded,
we need to detect user *inside* the hook contents. This will have
a lot code, but it works.

Please note that by removing IPM support from :admin, we will make
thing clear, not messy anymore. When being enabled, IPM can prevent
*admin* user from editting any image, but the native code allows
them to edit images (confused!).

I am not very sure why the previous Piwigo verions do not have this
problem. Maybe the template system has changed. Better means buggy