!bug: upload func. not work with Community 2.6
The community plugin (2.6) has a new setting to know whether
its function is enabled (for a user?). This setting seems to
be turnt off by default (FIXME: how to turn it on?), and that
will hide the "Upload link". Because IPM has its link displayed
by Community plugin, the upload func. seems not to work